refactor : Referee 클래스 패키지 이동
Referee에는 Cars를 움직이게 할 책임, 우승자를 판단할 책임이 있기 때문에 controller 패키지에서 model 패키지로 이동시켰습니다.
