+++ src/test/java/racingcar/model/RefereeTest.java
@@ -4,7 +4,6 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.List;
 import org.junit.jupiter.api.Test;
-import racingcar.controller.Referee;
 
 class RefereeTest {
     @Test
